requirements-docker.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
lxml==4.6.3
multitasking==0.0.9
numpy~=1.19.2
jupyter-dash==0.4.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
yfinance==0.1.62
pandas==1.2.4
matplotlib==3.4.2
FinNews==1.1.0
dash==1.20.0
plotly==5.1.0
cufflinks==0.17.3
boto3==1.17.98
mplfinance==0.12.7a17
yahoo-fin==0.8.9.1
scikit-learn == 0.24.2
tqdm==4.61.1
pandas-market-calendars==2.0
dash-bootstrap-components==0.13.0
SQLAlchemy==1.4.22
psycopg2-binary~=2.8.6
statsmodels==0.12.2
